requirements.txt
beautifulsoup4
requests
splinter
selenium
python-telegram-bot
